requirements.txt
fastapi[all]==0.115.6
uvicorn[standard]==0.22.*
python-dotenv
sqlalchemy==1.4.36
pymysql
mysql-connector-python
google-cloud-aiplatform
pymongo